backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
yfinance
pandas
numpy
python-dotenv
httpx
pydantic
feedparser
apscheduler
websockets